Update comments, cleanup
