+++ datafusion/physical-plan/src/aggregates/mod.rs
@@ -268,13 +268,6 @@ impl From<StreamType> for SendableRecordBatchStream {
     }
 }
 
-pub struct AggregateGroup {
-    aggregate_expressions: Vec<Vec<Arc<dyn PhysicalExpr>>>,
-    filter_expressions: Vec<Option<Arc<dyn PhysicalExpr>>>,
-    accumulators: Vec<AccumulatorItem>,
-    requirement: LexOrdering,
-}
-
 #[derive(Clone, Debug, PartialEq)]
 pub struct AggregateExprGroup {
     /// Aggregate expressions indices
@@ -990,24 +983,6 @@ fn create_accumulators(
         .iter()
         .map(|expr| expr.create_accumulator())
         .collect::<Result<Vec<_>>>()
-}
-
-/// returns a vector of ArrayRefs, where each entry corresponds to either the
-/// final value (mode = Final, FinalPartitioned and Single) or states (mode = Partial)
-fn finalize_aggregation_groups(
-    // accumulators: &[AccumulatorItem],
-    aggregate_groups: &[AggregateGroup],
-    // TODO: Use Mapping indices
-    mode: &AggregateMode,
-) -> Result<Vec<ArrayRef>> {
-    let elems = aggregate_groups
-        .iter()
-        .map(|aggregate_group| finalize_aggregation(&aggregate_group.accumulators, mode))
-        .collect::<Result<Vec<_>>>()?;
-    // TODO: Add proper indices
-    // Convert Vec<Vec<ArrayRef>> to Vec<ArrayRef>.
-    let res = elems.into_iter().flatten().collect::<Vec<_>>();
-    Ok(res)
 }
 
 /// returns a vector of ArrayRefs, where each entry corresponds to either the
